Move media data to common module
