+++ buildSrc/src/main/kotlin/extensions/DependencyHandlerExtensions.kt
@@ -20,7 +20,6 @@ import dependencies.TestAndroidDependencies
 import dependencies.TestDependencies
 import org.gradle.api.artifacts.Dependency
 import org.gradle.api.artifacts.dsl.DependencyHandler
-import org.gradle.kotlin.dsl.project
 
 /**
  * Adds a dependency to the `debugImplementation` configuration.
