+++ common/lessons-data/build.gradle.kts
@@ -23,7 +23,13 @@ import dependencies.Dependencies
 import dependencies.Dependencies.Firebase
 import dependencies.Dependencies.Kotlin
 import dependencies.Dependencies.Hilt
+import dependencies.Dependencies.Square.Moshi
+import dependencies.Dependencies.Square.Okhttp
+import dependencies.Dependencies.Square.Retrofit
 import extensions.addTestsDependencies
+import extensions.api
+import extensions.implementation
+import extensions.kapt
 
 plugins {
     id(BuildPlugins.Android.LIBRARY)
@@ -52,6 +58,7 @@ android {
 
 dependencies {
     implementation(project(BuildModules.Common.CORE))
+    implementation(project(BuildModules.Common.STORAGE))
 
     implementation(Kotlin.COROUTINES)
     implementation(Kotlin.COROUTINES_ANDROID)
@@ -64,5 +71,15 @@ dependencies {
     implementation(Firebase.DATABASE)
     implementation(Firebase.AUTH)
 
+    implementation(Moshi.kotlin)
+    kapt(Moshi.codegen)
+
+    api(Okhttp.okhttp)
+    implementation(Okhttp.logging)
+
+    implementation(Retrofit.retrofit2)
+    implementation(Retrofit.moshiConverter)
+
     addTestsDependencies()
+    testImplementation(project(BuildModules.Libraries.TEST_UTILS))
 }
