+++ common/lessons-data/src/main/java/app/ss/lessons/data/model/api/SSAudio.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021. Adventech <info@adventech.io>
+ * Copyright (c) 2022. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,9 +20,20 @@
  * THE SOFTWARE.
  */
 
-package app.ss.media.model.request
+package app.ss.lessons.data.model.api
 
-internal data class SSMediaRequest(
-    val language: String,
-    val quarterlyId: String
+import androidx.annotation.Keep
+import com.squareup.moshi.JsonClass
+
+@Keep
+@JsonClass(generateAdapter = true)
+data class SSAudio(
+    val id: String,
+    val artist: String,
+    val image: String,
+    val imageRatio: String,
+    val src: String,
+    val target: String,
+    val targetIndex: String,
+    val title: String
 )
