+++ common/lessons-data/src/main/java/app/ss/lessons/data/model/media/AudioFile.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021. Adventech <info@adventech.io>
+ * Copyright (c) 2022. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,25 +20,32 @@
  * THE SOFTWARE.
  */
 
-package app.ss.media.api
+package app.ss.lessons.data.model.media
 
-import app.ss.media.model.SSAudio
-import app.ss.media.model.SSVideosInfo
-import retrofit2.Response
-import retrofit2.http.GET
-import retrofit2.http.Path
+import android.net.Uri
+import androidx.annotation.Keep
+import app.ss.storage.db.entity.AudioFileEntity
 
-interface SSMediaApi {
+@Keep
+data class AudioFile(
+    val id: String,
+    val title: String = "",
+    val artist: String = "",
+    val source: Uri = Uri.EMPTY,
+    val image: String = "",
+    val imageRatio: String = "",
+    val target: String = "",
+    val targetIndex: String = "",
+    val duration: Long = 0
+)
 
-    @GET("api/v1/{lang}/quarterlies/{quarterly_id}/audio.json")
-    suspend fun getAudio(
-        @Path("lang") language: String,
-        @Path("quarterly_id") quarterlyId: String
-    ): Response<List<SSAudio>>
-
-    @GET("api/v1/{lang}/quarterlies/{quarterly_id}/video.json")
-    suspend fun getVideo(
-        @Path("lang") language: String,
-        @Path("quarterly_id") quarterlyId: String
-    ): Response<List<SSVideosInfo>>
-}
+fun AudioFileEntity.toAudio(): AudioFile = AudioFile(
+    id = id,
+    artist = artist,
+    image = image,
+    imageRatio = imageRatio,
+    source = Uri.parse(src),
+    target = target,
+    targetIndex = targetIndex,
+    title = title,
+)
