+++ common/lessons-data/src/main/java/app/ss/lessons/data/repository/media/SSMediaRepository.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021. Adventech <info@adventech.io>
+ * Copyright (c) 2022. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,16 +20,16 @@
  * THE SOFTWARE.
  */
 
-package app.ss.media.repository
+package app.ss.lessons.data.repository.media
 
-import app.ss.media.api.SSMediaApi
-import app.ss.media.model.SSAudio
-import app.ss.media.model.SSVideosInfo
-import app.ss.media.model.request.SSMediaRequest
-import app.ss.media.model.toAudio
-import app.ss.media.model.toEntity
-import app.ss.media.playback.model.AudioFile
+import app.ss.lessons.data.api.SSMediaApi
+import app.ss.lessons.data.model.api.SSAudio
+import app.ss.lessons.data.model.api.SSVideosInfo
+import app.ss.lessons.data.model.api.request.SSMediaRequest
+import app.ss.lessons.data.model.media.AudioFile
+import app.ss.lessons.data.model.media.toAudio
 import app.ss.storage.db.dao.AudioDao
+import app.ss.storage.db.entity.AudioFileEntity
 import com.cryart.sabbathschool.core.extensions.coroutines.SchedulerProvider
 import com.cryart.sabbathschool.core.response.Resource
 import kotlinx.coroutines.withContext
@@ -117,3 +117,14 @@ private fun String.toMediaRequest(): SSMediaRequest? {
     val id = langId.substringAfter('-')
     return SSMediaRequest(lang, id)
 }
+
+fun SSAudio.toEntity(): AudioFileEntity = AudioFileEntity(
+    id = id,
+    artist = artist,
+    image = image,
+    imageRatio = imageRatio,
+    src = src,
+    target = target,
+    targetIndex = targetIndex,
+    title = title,
+)
