+++ features/media/src/main/kotlin/app/ss/media/playback/service/MusicService.kt
@@ -5,11 +5,12 @@ import android.os.Bundle
 import android.support.v4.media.MediaBrowserCompat
 import androidx.media.MediaBrowserServiceCompat
 import androidx.media.session.MediaButtonReceiver
-import app.ss.media.playback.MediaNotifications
+import app.ss.lessons.data.repository.media.SSMediaRepository
+import app.ss.media.playback.BACKWARD
 import app.ss.media.playback.FORWARD
+import app.ss.media.playback.MediaNotifications
 import app.ss.media.playback.NOTIFICATION_ID
 import app.ss.media.playback.PLAY_PAUSE
-import app.ss.media.playback.BACKWARD
 import app.ss.media.playback.STOP_PLAYBACK
 import app.ss.media.playback.extensions.isIdle
 import app.ss.media.playback.extensions.playPause
@@ -18,7 +19,6 @@ import app.ss.media.playback.model.MediaId.Companion.CALLER_OTHER
 import app.ss.media.playback.model.MediaId.Companion.CALLER_SELF
 import app.ss.media.playback.players.SSAudioPlayer
 import app.ss.media.playback.receivers.BecomingNoisyReceiver
-import app.ss.media.repository.SSMediaRepository
 import com.cryart.sabbathschool.core.extensions.coroutines.SchedulerProvider
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
