+++ features/media/src/main/kotlin/app/ss/media/playback/ui/nowPlaying/NowPlayingViewModel.kt
@@ -25,14 +25,14 @@ package app.ss.media.playback.ui.nowPlaying
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import app.ss.lessons.data.model.media.AudioFile
+import app.ss.lessons.data.repository.media.SSMediaRepository
 import app.ss.media.playback.AudioQueueManager
 import app.ss.media.playback.PlaybackConnection
 import app.ss.media.playback.UPDATE_META_DATA
 import app.ss.media.playback.extensions.id
 import app.ss.media.playback.extensions.isPlaying
-import app.ss.media.playback.model.AudioFile
 import app.ss.media.playback.model.toAudio
-import app.ss.media.repository.SSMediaRepository
 import com.cryart.sabbathschool.core.extensions.coroutines.flow.stateIn
 import com.cryart.sabbathschool.core.extensions.intent.lessonIndex
 import com.cryart.sabbathschool.core.extensions.intent.readIndex
