+++ features/media/src/main/kotlin/app/ss/media/playback/ui/video/VideoListViewModel.kt
@@ -25,8 +25,8 @@ package app.ss.media.playback.ui.video
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.ss.media.model.SSVideosInfo
-import app.ss.media.repository.SSMediaRepository
+import app.ss.lessons.data.model.api.SSVideosInfo
+import app.ss.lessons.data.repository.media.SSMediaRepository
 import com.cryart.sabbathschool.core.extensions.coroutines.SchedulerProvider
 import com.cryart.sabbathschool.core.extensions.coroutines.flow.stateIn
 import com.cryart.sabbathschool.core.extensions.intent.lessonIndex
