+++ features/pdf/src/main/kotlin/app/ss/pdf/PdfReader.kt
@@ -27,7 +27,7 @@ import android.content.Intent
 import android.net.Uri
 import androidx.annotation.Keep
 import app.ss.lessons.data.model.LessonPdf
-import app.ss.media.model.MediaAvailability
+import app.ss.lessons.data.model.media.MediaAvailability
 import app.ss.pdf.ui.ARG_MEDIA_AVAILABILITY
 import app.ss.pdf.ui.ARG_PDF_FILES
 import app.ss.pdf.ui.SSReadPdfActivity
